Update cpp_magic.h
I might be using an adaptation of this header for other purposes and I found an scenario that failed for me. I just wanted to contribute back the fix, even though it isn't needed in your project.

As the "HAS_ARGS" macro is before this modification, a substitution like the next:

"HAS_ARGS ((int) 0, (int) 0)"

Fails to compile, as the expansion would substitute to "END_OF_ARGUMENTS (int) 0", which is an invocation with one token.

The easiest fix I found is to just break the token name in two and append the tail after getting the first variadic argument.